seperated optimizers, going to add ResNet generator network
